backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
djangorestframework==3.16.1
django-filter==24.3
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1
reportlab==4.2.5
WeasyPrint==62.3
matplotlib==3.9.2
sqlparse==0.5.3
tzdata==2025.2
# Driver para PostgreSQL
psycopg2-binary
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.3.1
gunicorn
